app/requirements.txt
langchain==0.3.24
langchain-ollama==0.3.2
langchain-tavily==0.1.5
langgraph==0.3.32
streamlit==1.44.1
langchain-groq==0.3.2
python-dotenv